+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+- Allow returning result without throwing if it has any command error
+- Add ```execWithModel``` method to convert an array result to custom model
+
 ## 1.0.1
 
 - Reformat